requirements.txt
absl-py==2.3.1
asttokens==3.0.0
autocommand==2.2.2
backports.tarfile==1.2.0
bypy==1.8.9
certifi==2025.8.3
charset-normalizer==3.4.3
clip==1.0
cloudpickle==3.1.1
cmake==4.1.0
decorator==5.2.1
diffusers==0.26.0
dill==0.4.0
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.19.1
fsspec==2025.7.0
ftfy==6.3.1
grpcio==1.74.0
hf-xet==1.1.9
huggingface-hub==0.20.2
idna==3.10
importlib_metadata==8.0.0
inflect==7.3.1
ipdb==0.13.13
ipython==8.37.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.2
Jinja2==3.1.6
lit==18.1.8
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.3.0
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.4.2
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
packaging==24.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.52
protobuf==6.32.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scipy==1.15.3
stack-data==0.6.3
submitit==1.5.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
timm==1.0.19
tomli==2.0.1
torch==2.0.1
torch-fidelity==0.3.0
torchaudio==2.0.2
torchdiffeq==0.2.4
torchvision==0.15.2
tqdm==4.67.1
traitlets==5.14.3
triton==2.0.0
typeguard==4.3.0
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.19.2
